All codes/CHATBOT/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.11.1",
    "openai-agents>=0.17.0",
    "python-dotenv>=1.2.2",
]
